pyproject.toml
[tool.poetry]
name = "selfdiffusion"
version = "0.1.0"
description = ""
authors = ["Jonathan Pelletier <jonathan.pelletier-aafz@thecloudco.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.0.1"
halo = "^0.0.31"
pillow = "^9.5.0"
face-recognition = "^1.3.0"
ultralytics = "^8.0.66"

[tool.poetry.scripts]
selfdiffusion = 'selfdiffusion.cli:main'


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
